implement project-level access control and synchronization from Keycloak groups

+++ src/main/java/de/openfabtwin/controllers/CommentController.java
@@ -34,7 +34,10 @@ public class CommentController implements CommentsApi {
     @Override
     public ResponseEntity<CommentGET> createComment(String version, String projectId, String topicId, CommentPOST commentPOST) {
         props.validateVersion(version);
-        //TODO: validate user has permission to project
+        boolean hasAccess = securityContextService.hasProjectAccess(projectId);
+        if (!hasAccess) {
+            throw new AccessDeniedException("User does not have access to project");
+        }
         UserRole role = securityContextService.getCurrentUserRole();
         authorizationService.assertCan(role, Actions.Topic.CREATE_COMMENT);
         CommentEntity created = commentService.create(projectId, topicId, commentPOST);
@@ -46,7 +49,10 @@ public class CommentController implements CommentsApi {
     @Override
     public ResponseEntity<Void> deleteComment(String version, String projectId, String topicId, String commentId) {
         props.validateVersion(version);
-        //TODO: validate user has permission to project
+        boolean hasAccess = securityContextService.hasProjectAccess(projectId);
+        if (!hasAccess) {
+            throw new AccessDeniedException("User does not have access to project");
+        }
         UserRole role = securityContextService.getCurrentUserRole();
         authorizationService.assertCan(role, Actions.Comment.DELETE);
         commentService.delete(commentId, topicId, projectId);
@@ -56,7 +62,10 @@ public class CommentController implements CommentsApi {
     @Override
     public ResponseEntity<CommentGET> getCommentById(String version, String projectId, String topicId, String commentId) {
         props.validateVersion(version);
-        //TODO: validate user has permission to project
+        boolean hasAccess = securityContextService.hasProjectAccess(projectId);
+        if (!hasAccess) {
+            throw new AccessDeniedException("User does not have access to project");
+        }
         UserRole role = securityContextService.getCurrentUserRole();
         CommentEntity comment = commentService.getById(commentId, topicId, projectId);
         CommentGET dto = commentMapper.toDto(comment, topicId);
@@ -67,7 +76,10 @@ public class CommentController implements CommentsApi {
     @Override
     public ResponseEntity<List<CommentGET>> getTopicComment(String version, String projectId, String topicId, String $filter, String $orderby) {
         props.validateVersion(version);
-        //TODO: validate user has permission to project
+        boolean hasAccess = securityContextService.hasProjectAccess(projectId);
+        if (!hasAccess) {
+            throw new AccessDeniedException("User does not have access to project");
+        }
         UserRole role = securityContextService.getCurrentUserRole();
         List<CommentGET> comments = commentService.getAll(projectId, topicId, $filter, $orderby)
                 .stream()
@@ -83,7 +95,10 @@ public class CommentController implements CommentsApi {
     @Override
     public ResponseEntity<CommentGET> updateComment(String version, String projectId, String topicId, String commentId, CommentPUT commentPUT) {
         props.validateVersion(version);
-        //TODO: validate user has permission to project
+        boolean hasAccess = securityContextService.hasProjectAccess(projectId);
+        if (!hasAccess) {
+            throw new AccessDeniedException("User does not have access to project");
+        }
         UserRole role = securityContextService.getCurrentUserRole();
         authorizationService.assertCan(role, Actions.Comment.UPDATE);
         CommentEntity updated = commentService.update(commentId, topicId, projectId, commentPUT);
